Score AI as average points above benchmark

+++ test/test_evolution_algs.cpp
@@ -6,16 +6,16 @@
 
 using namespace std;
 
-extern vector<int> populationDropN(const unordered_map<int, int>& scores, unsigned int numDrop);
+extern vector<int> populationDropN(const unordered_map<int, float>& scores, unsigned int numDrop);
 
 TEST(TestEvolutionAlgs, CanDropNoNumers) {
-    unordered_map<int, int> scores = {{0, 100}, {1, 50}, {2, 150}, {3, 75}};
+    unordered_map<int, float> scores = {{0, 100}, {1, 50}, {2, 150}, {3, 75}};
     auto result = populationDropN(scores, 0);
     ASSERT_EQ((vector<int>{2, 0, 3, 1}), result);
 }
 
 TEST(TestEvolutionAlgs, CanDrop2SmallestNumbers) {
-    unordered_map<int, int> scores = {{0, 100}, {1, 50}, {2, 150}, {3, 75}};
+    unordered_map<int, float> scores = {{0, 100}, {1, 50}, {2, 150}, {3, 75}};
     auto result = populationDropN(scores, 2);
 
     ASSERT_EQ((vector<int>{2, 0}), result);
